test(core): move custom ltc address tests to utxo/custom/ltc.ts
Issue: BG-38773
